requirements.txt
pandas 
numpy
matplotlib
cv2
os
streamlit
PyPDF2
transformer
